clears the selection in the tickets grid when active filters change
The selection in <vaadin-grid> works in a way that when the grid contents changes the selected rows (say 1, 2 and 5) would remain selected even if their contents has changed. For example one extra item is added to the grid and it appears at the row 2. The remaining rows are shifted down, and what was on the row 2 now appears on the row 3. The selection is still at the rows 1, 2 and 5 but the contents of the selected rows is now different.

For that reason it is more confusing than meaningful to keep the selection when the filters change.